Refactor deployment workflow and update environment variables
The deployment workflow was refactored to update the branch name from `main` to `master`. Additionally, the testing environment now uses a Node.js 16 container image instead of the previous default image. The MongoDB connection configuration was also modified to use a different connection protocol and cluster address, as well as new default credentials. Lastly, a MongoDB service was added to the testing environment using the `mongo` image.

.gitignore was updated to exclude `.vscode` directory.